refactor recon.tracking.seedtracker
