+++ tracking/src/main/java/org/hps/recon/tracking/TrackerReconDriver.java
@@ -12,6 +12,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.logging.Logger;
 
+import org.hps.recon.tracking.seedtracker.SeedTracker;
 import org.lcsim.event.EventHeader;
 import org.lcsim.event.RelationalTable;
 import org.lcsim.event.Track;
@@ -37,47 +38,47 @@ public final class TrackerReconDriver extends Driver {
     private static final Logger LOGGER = Logger.getLogger(TrackerReconDriver.class.getPackage().getName());
 
     private String subdetectorName = "Tracker";
-    
+
     // Debug flag.
     private boolean debug = false;
-    
+
     // Tracks found across all events.
     int ntracks = 0;
-    
+
     // Number of events processed.
     int nevents = 0;
-    
+
     // Cache detector object.
     Detector detector = null;
-    
+
     // Default B-field value.
     private double bfield = 0.5;
-    
+
     // Tracking strategies resource path.
     private String strategyResource = "HPS-Test-4pt1.xml";
-    
+
     // Output track collection.
     private String trackCollectionName = "MatchedTracks";
-    
+
     // HelicalTrackHit input collection.
     private String stInputCollectionName = "RotatedHelicalTrackHits";
-    
+
     // Include MS (done by removing XPlanes from the material manager results)
     private boolean includeMS = true;
-    
+
     // number of repetitive fits on confirmed/extended tracks
     private int _iterativeConfirmed = 3;
-    
+
     // use HPS implementation of material manager
     private boolean _useHPSMaterialManager = true;
-    
+
     // enable the use of sectoring using sector binning in SeedTracker
     private boolean _applySectorBinning = true;
-    
+
     private double rmsTimeCut = -1;
-    
+
     private boolean rejectUncorrectedHits = true;
-    
+
     private boolean rejectSharedHits = false;
 
     public TrackerReconDriver() {
@@ -86,7 +87,7 @@ public final class TrackerReconDriver extends Driver {
     public void setDebug(boolean debug) {
         this.debug = debug;
     }
-    
+
     public void setSubdetectorName(String subdetectorName) {
         this.subdetectorName = subdetectorName;
     }
@@ -203,7 +204,7 @@ public final class TrackerReconDriver extends Driver {
             HitTimeTrackCheck timeCheck = new HitTimeTrackCheck(rmsTimeCut);
             timeCheck.setDebug(debug);
             stFinal.setTrackCheck(timeCheck);
-        }        
+        }
     }
 
     /**
@@ -239,8 +240,7 @@ public final class TrackerReconDriver extends Driver {
 
         if (rejectUncorrectedHits) {
             Iterator<Track> iter = tracks.iterator();
-            trackLoop:
-            while (iter.hasNext()) {
+            trackLoop: while (iter.hasNext()) {
                 Track track = iter.next();
                 for (TrackerHit hit : track.getTrackerHits()) {
                     HelicalTrackHit hth = (HelicalTrackHit) hit;
@@ -274,8 +274,7 @@ public final class TrackerReconDriver extends Driver {
                 }
             }
             Iterator<Track> iter = tracks.iterator();
-            trackLoop:
-            while (iter.hasNext()) {
+            trackLoop: while (iter.hasNext()) {
                 Track track = iter.next();
                 for (TrackerHit hit : track.getTrackerHits()) {
                     Collection<TrackerHit> htsList = hitToStrips.allFrom(hitToRotated.from(hit));
