+++ tracking/src/main/java/org/hps/recon/tracking/gbl/GBLOutput.java
@@ -18,10 +18,10 @@ import java.util.Map;
 import org.apache.commons.math.geometry.Rotation;
 import org.apache.commons.math.geometry.Vector3D;
 import org.hps.recon.tracking.CoordinateTransformations;
-import org.hps.recon.tracking.MaterialSupervisor;
-import org.hps.recon.tracking.MultipleScattering;
-import org.hps.recon.tracking.MultipleScattering.ScatterPoint;
-import org.hps.recon.tracking.MultipleScattering.ScatterPoints;
+import org.hps.recon.tracking.seedtracker.MaterialSupervisor;
+import org.hps.recon.tracking.seedtracker.MultipleScattering;
+import org.hps.recon.tracking.seedtracker.MultipleScattering.ScatterPoint;
+import org.hps.recon.tracking.seedtracker.MultipleScattering.ScatterPoints;
 import org.hps.recon.tracking.TrackUtils;
 import org.hps.recon.tracking.TrackerHitUtils;
 import org.lcsim.detector.IDetectorElement;
