+++ tracking/src/main/java/org/hps/recon/tracking/gbl/GBLRefitterDriver.java
@@ -7,8 +7,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import org.apache.commons.math3.util.Pair;
-import org.hps.recon.tracking.MaterialSupervisor;
-import org.hps.recon.tracking.MultipleScattering;
+import org.hps.recon.tracking.seedtracker.MaterialSupervisor;
+import org.hps.recon.tracking.seedtracker.MultipleScattering;
 import org.hps.recon.tracking.TrackUtils;
 import org.lcsim.detector.DetectorElementStore;
 import org.lcsim.detector.IDetectorElement;
